fix a bunch of stuff the rebase broke
